backend/requirements/base.txt
Django>=1.8.5
django-environ>=0.3.0
django-admin-bootstrapped>=2.4.0
djangorestframework>=3.2.3
django-allauth>=0.23.0
django-rest-auth>=0.5.0
drf-nested-routers>=0.10.0
pillow>=3.0.0